feat: Move HomeScreen to home/Home.kt and Rename MainViewModel to HomeViewModel
